requirements.txt
google-cloud-language
google-cloud-storage
textgenrnn
tensorflow